dom/tests/mochitest/bugs/test_bug664737.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=664737
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 664737</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=664737">Mozilla Bug 664737</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 664737 **/
function isUndefined(name) {
    try {
        var rv = history[name];
        is(rv, undefined, "No value found for history." + name);
    } catch (e) {
        ok(false, "Accessing history." + name + " threw: " + e);
    }
}

isUndefined("current");
isUndefined("previous");
isUndefined("next");


</script>
</pre>
</body>
</html>
